Add files via upload
First working version of the Gita page generator. It reads the texts and commentaries from a  google spreadsheet. It uses a sample tagged file for wiki presentation and a manually computed index of the sample file. The tags are replaced with text read from the google sheet with appropriate page header  etc. A file for each sloka is generated and stored in ./verses folder.